demo/frontend/requirements.txt
streamlit>=0.89.0
requests>=2.26.0
pandas>=1.3.0
python-dotenv>=0.19.0
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2